+++ docs/.vitepress/zh.ts
@@ -192,6 +192,8 @@ function sidebarIs(): DefaultTheme.SidebarItem[] {
       base: 'func/misc/',
       collapsed: false,
       items: [
+        { text: 'arrayToCSV\n二维数组转 CSV', link: 'arrayToCSV' },
+        { text: 'objectListToCSV\n对象列表转 CSV', link: 'objectListToCSV' },
         { text: 'acceptableFileName\n检查拓展名', link: 'acceptableFileName' },
         { text: 'acceptableFileType\n检查 MIME 类型', link: 'acceptableFileType' },
         { text: 'getAcceptableExtByMIME\n获取拓展名', link: 'getAcceptableExtByMIME' },